Added AF choice side for dorm management
Added the air force side to be able to make that a choice. It still goes to the same choices for phone numbers but the AF side is there. the bug with the iphone not being able to sync numbers due to not having the phone app is still there