Parsed data
currently it reads into great life hawaii and is able to pull the names of the events on the front page into events list. 

Things that are missing, storing the data with shared preferences, comparing the data with shared preferences and giving a notification of a new event when it changes.