Test - Timer and comparision
In the test page I am able to to run a timer ranging from 15 seconds - X days. This timer will run the any function once the time is up.  This allows for updating the list as needed

Comparison. I am able to compare the old array list and the new array list and replace the old one with the new

next task:
possibly need a counter to be able to not notify on the first comparison.

Need to set up local notification.